+++ lihil/signature/signature.py
@@ -8,7 +8,14 @@ from lihil.config import AppConfig
 from lihil.interface import Base, IEncoder, Record
 from lihil.problems import ValidationProblem
 from lihil.utils.string import find_path_keys
-from lihil.vendors import FormData, Headers, QueryParams, Request, cookie_parser, WebSocket
+from lihil.vendors import (
+    FormData,
+    Headers,
+    QueryParams,
+    Request,
+    WebSocket,
+    cookie_parser,
+)
 
 from .params import (
     BodyParam,
@@ -154,6 +161,11 @@ class EndpointSignature[R](Base):
             )
         )
 
+    @property
+    def media_type(self) -> str:
+        default = "application/json"
+        return next(iter(self.return_params.values())).content_type or default
+
     @classmethod
     def from_function[FR](
         cls,
@@ -163,9 +175,10 @@ class EndpointSignature[R](Base):
         app_config: AppConfig | None = None,
     ) -> "EndpointSignature[FR]":
         path_keys = find_path_keys(route_path)
-
+        # Rename ParmaParser to FuncParser
         parser = ParamParser(graph, path_keys, app_config=app_config)
-        params = parser.parse(f, path_keys)
+        params = parser.parse(f)
+        # TODO: let ParamParser parse returns too
         return_params = parse_returns(
             signature(f).return_annotation, app_config=app_config
         )