+++ tests/test_plugins/test_bus.py
@@ -1,4 +1,5 @@
 from typing import Any
+import pytest
 
 from lihil import Resp, Route, status
 from lihil.plugins.bus import Event, EventBus
@@ -21,10 +22,12 @@ async def listen_twice(created: TodoCreated, _: Any):
     assert created.content
 
 
-bus_route = Route("/bus", listeners=[listen_create, listen_twice])
+@pytest.fixture
+def bus_route():
+    return Route("/bus", listeners=[listen_create, listen_twice])
 
 
-async def test_bus_is_singleton():
+async def test_bus_is_singleton(bus_route: Route):
     async def create_todo(
         name: str, content: str, bus: EventBus
     ) -> Resp[None, status.OK]:
@@ -38,7 +41,13 @@ async def test_bus_is_singleton():
     assert any(p.type_ is EventBus for p in ep.sig.plugins.values())
 
 
-async def test_call_ep_invoke_bus():
+async def test_call_ep_invoke_bus(bus_route: Route):
+    async def create_todo(
+        name: str, content: str, bus: EventBus
+    ) -> Resp[None, status.OK]:
+        await bus.publish(TodoCreated(name, content))
+
+    bus_route.post(create_todo)
     ep = bus_route.get_endpoint("POST")
     client = LocalClient()
     await client.call_endpoint(ep, query_params=dict(name="1", content="2"))