+++ CHANGELOG.md
@@ -913,5 +913,39 @@ NOTE that for this to work, both `ws_handler` and `ws_factory` should name `WebS
 
 ## version 0.2.4
 
+A simple, daily maintainence patch, mainly for refactoring.
 
-- [ ] static response
+### Improvements
+
+- [x] graph ignore lihil primitives(Request, Websocket, ...) by default
+
+- [x] static response
+
+now static response with non-generator returns would return `StaticResponse` instead.
+
+### Refactors
+
+- [x] No longer cache Route by path.
+
+previously, instances of `Route` will be cached by their path
+
+```python
+assert Route("user") is Route("user")
+```
+
+This  was for the sake of convenience, so that user do
+```python
+@Route("/user").get
+def get_user(): ...
+
+@Route("/user").post
+def create_user(): ...
+```
+
+But the fundamental flaws of this design is that:
+
+1. users might not expect this.
+2. testing is harder.
+
+
+- [x] specialized param meta